zng_ext_input::focus

Module iter

Focusable info tree iterators.

Structs

FocusTreeFilterIter

An iterator that filters a focusable widget tree.

FocusTreeIter

Iterator over all focusable items in a branch of the widget tree.

IterFocusable

Filter a widget info iterator to only focusable items.

Traits

IterFocusableExt

Filter-maps an iterator of WidgetInfo to WidgetFocusInfo.