Struct zng_ext_input::focus::FocusChangedArgs

pub struct FocusChangedArgs {
    pub timestamp: DInstant,
    pub prev_focus: Option<InteractionPath>,
    pub new_focus: Option<InteractionPath>,
    pub highlight: bool,
    pub cause: FocusChangedCause,
    pub enabled_nav: FocusNavAction,
    /* private fields */
}

FOCUS_CHANGED_EVENT arguments.

Fields
timestamp: DInstant

Instant the event happened.

prev_focus: Option<InteractionPath>

Previously focused widget.

new_focus: Option<InteractionPath>

Newly focused widget.

highlight: bool

If the focused widget should visually indicate that it is focused.

This is true when the focus change is caused by a key press, false when it is caused by a mouse click.

Some widgets, like text input, may ignore this field and always indicate that they are focused.

cause: FocusChangedCause

What caused this event.

enabled_nav: FocusNavAction

Focus navigation actions that can move the focus away from the new_focus.

Implementations

impl FocusChangedArgs

pub fn new( timestamp: impl Into<DInstant>, propagation_handle: EventPropagationHandle, prev_focus: impl Into<Option<InteractionPath>>, new_focus: impl Into<Option<InteractionPath>>, highlight: impl Into<bool>, cause: impl Into<FocusChangedCause>, enabled_nav: impl Into<FocusNavAction>, ) -> Self

New args from values that convert into the argument types.

pub fn now( prev_focus: impl Into<Option<InteractionPath>>, new_focus: impl Into<Option<InteractionPath>>, highlight: impl Into<bool>, cause: impl Into<FocusChangedCause>, enabled_nav: impl Into<FocusNavAction>, ) -> Self

Arguments for event that happened now (INSTANT.now).

impl FocusChangedArgs

pub fn is_widget_move(&self) -> bool

If the focus is still in the same widget, but the widget path changed.

pub fn is_enabled_change(&self) -> bool

If the focus is still in the same widget path, but some or all interactivity has changed.

pub fn is_highlight_changed(&self) -> bool

If the focus is still in the same widget but highlight changed.

pub fn is_focus(&self, widget_id: WidgetId) -> bool

If widget_id is the new focus and was not before.

pub fn is_blur(&self, widget_id: WidgetId) -> bool

If widget_id is the previous focus and is not now.

pub fn is_focus_enter(&self, widget_id: WidgetId) -> bool

If widget_id is the new focus or a parent of the new focus and was not a parent of the previous focus.

pub fn is_focus_leave(&self, widget_id: WidgetId) -> bool

If widget_id is the previous focus or a parent of the previous focus and is not a parent of the new focus.

pub fn is_focused(&self, widget_id: WidgetId) -> bool

If the widget is the new focus.

pub fn is_focus_within(&self, widget_id: WidgetId) -> bool

If the widget is in the new focus path.