Struct zng_ext_input::focus::WidgetFocusInfo

pub struct WidgetFocusInfo { /* private fields */ }

WidgetInfo wrapper that adds focus information for each widget.

Implementations

impl WidgetFocusInfo

pub fn new( widget_info: WidgetInfo, focus_disabled_widgets: bool, focus_hidden_widgets: bool ) -> Self

Wrap a widget_info reference to enable focus info querying.

See the FOCUS.focus_disabled_widgets and FOCUS.focus_hidden_widgets config for more details on the parameters.

pub fn info(&self) -> &WidgetInfo

Full widget info.

pub fn focus_disabled_widgets(&self) -> bool

If DISABLED widgets are focusable in this tree.

See the FOCUS.focus_disabled_widgets config for more details.

pub fn focus_hidden_widgets(&self) -> bool

If Hidden widgets are focusable in this tree.

See the FOCUS.focus_hidden_widgets config for more details.

pub fn root(&self) -> Self

Root focusable.

pub fn focus_tree(&self) -> FocusInfoTree

Clone a reference to the FocusInfoTree that owns this widget.

pub fn is_focusable(&self) -> bool

If the widget is focusable.

Note

This is probably true, the only way to get a WidgetFocusInfo for a non-focusable widget is by calling into_focus_info or explicitly constructing one.

Focus scopes are also focusable.

pub fn is_scope(&self) -> bool

Is focus scope.

pub fn is_alt_scope(&self) -> bool

Is ALT focus scope.

pub fn focus_info(&self) -> FocusInfo

Widget focus metadata.

pub fn focus_info_ignore_blocked(&self) -> FocusInfo

Widget focus metadata, all things equal except the widget interactivity is blocked.

pub fn ancestors(&self) -> impl Iterator<Item = WidgetFocusInfo>

Iterator over focusable parent -> grandparent -> .. -> root.

pub fn self_and_ancestors(&self) -> impl Iterator<Item = WidgetFocusInfo>

Iterator over self -> focusable parent -> grandparent -> .. -> root.

pub fn scopes(&self) -> impl Iterator<Item = WidgetFocusInfo>

Iterator over focus scopes parent -> grandparent -> .. -> root.

pub fn parent(&self) -> Option<WidgetFocusInfo>

Reference to the focusable parent that contains this widget.

pub fn scope(&self) -> Option<WidgetFocusInfo>

Reference the focus scope parent that contains the widget.

pub fn alt_scope(&self) -> Option<WidgetFocusInfo>

Reference the ALT focus scope closest with the current widget.

Closest Alt Scope

If self is already an ALT scope or is in one, moves to a sibling ALT scope, nested ALT scopes are ignored.

If self is a normal scope, moves to the first descendant ALT scope, otherwise..

Recursively searches for an ALT scope sibling up the scope tree.

pub fn in_alt_scope(&self) -> bool

Widget is in a ALT scope or is an ALT scope.

pub fn on_focus_scope_move<'f>( &self, last_focused: impl FnOnce(WidgetId) -> Option<&'f WidgetPath>, reverse: bool ) -> Option<WidgetFocusInfo>

Widget the focus needs to move to when self gets focused.

Input

last_focused: A function that returns the last focused widget within a focus scope identified by WidgetId.

reverse: If the focus is reversing into self.

Returns

Returns the different widget the focus must move to after focusing in self that is a focus scope.

If self is not a FocusScope always returns None.

pub fn descendants(&self) -> FocusTreeIter<TreeIter>

Iterator over the focusable widgets contained by this widget.

pub fn self_and_descendants(&self) -> FocusTreeIter<TreeIter>

Iterator over self and the focusable widgets contained by it.

pub fn has_tab_descendant(&self) -> bool

If the focusable has any focusable descendant that is not TabIndex::SKIP

pub fn first_tab_descendant(&self) -> Option<WidgetFocusInfo>

First descendant considering TAB index.

pub fn last_tab_descendant(&self) -> Option<WidgetFocusInfo>

Last descendant considering TAB index.

pub fn next_focusables(&self) -> FocusTreeIter<TreeIter>

Iterator over all focusable widgets in the same scope after this widget.

pub fn next_focusable(&self) -> Option<WidgetFocusInfo>

Next focusable in the same scope after this widget.

pub fn next_tab_focusable(&self, skip_self: bool) -> Option<WidgetFocusInfo>

Next focusable in the same scope after this widget respecting the TAB index.

If self is set to TabIndex::SKIP returns the next non-skip focusable in the same scope after this widget.

If skip_self is true, does not include widgets inside self.

pub fn prev_focusables(&self) -> FocusTreeIter<RevTreeIter>

Iterator over all focusable widgets in the same scope before this widget in reverse.

pub fn prev_focusable(&self) -> Option<WidgetFocusInfo>

Previous focusable in the same scope before this widget.

pub fn prev_tab_focusable(&self, skip_self: bool) -> Option<WidgetFocusInfo>

Previous focusable in the same scope after this widget respecting the TAB index.

If self is set to TabIndex::SKIP returns the previous non-skip focusable in the same scope before this widget.

If skip_self is true, does not include widgets inside self.

pub fn next_tab(&self, skip_self: bool) -> Option<WidgetFocusInfo>

Widget to focus when pressing TAB from this widget.

Set skip_self to not enter self, that is, the focus goes to the next sibling or next sibling descendant.

Returns None if the focus does not move to another widget.

pub fn prev_tab(&self, skip_self: bool) -> Option<WidgetFocusInfo>

Widget to focus when pressing SHIFT+TAB from this widget.

Set skip_self to not enter self, that is, the focus goes to the previous sibling or previous sibling descendant.

Returns None if the focus does not move to another widget.

pub fn nearest( &self, origin: PxPoint, max_radius: Px ) -> Option<WidgetFocusInfo>

Find the focusable descendant with center point nearest of origin within the max_radius.

pub fn nearest_filtered( &self, origin: PxPoint, max_radius: Px, filter: impl FnMut(WidgetFocusInfo) -> bool ) -> Option<WidgetFocusInfo>

Find the descendant with center point nearest of origin within the max_radius and approved by the filter closure.

pub fn nearest_bounded_filtered( &self, origin: PxPoint, max_radius: Px, bounds: PxRect, filter: impl FnMut(WidgetFocusInfo) -> bool ) -> Option<WidgetFocusInfo>

Find the descendant with center point nearest of origin within the max_radius and inside bounds; and approved by the filter closure.

pub fn nearest_oriented( &self, origin: PxPoint, max_distance: Px, orientation: Orientation2D ) -> Option<WidgetFocusInfo>

Find the focusable descendant with center point nearest of origin within the max_distance and with orientation to origin.

pub fn nearest_oriented_filtered( &self, origin: PxPoint, max_distance: Px, orientation: Orientation2D, filter: impl FnMut(WidgetFocusInfo) -> bool ) -> Option<WidgetFocusInfo>

Find the focusable descendant with center point nearest of origin within the max_distance and with orientation to origin that passes the filter.

pub fn focusable_up(&self) -> Option<WidgetFocusInfo>

Closest focusable in the same scope above this widget.

pub fn focusable_down(&self) -> Option<WidgetFocusInfo>

Closest focusable in the same scope below this widget.

pub fn focusable_left(&self) -> Option<WidgetFocusInfo>

Closest focusable in the same scope to the left of this widget.

pub fn focusable_right(&self) -> Option<WidgetFocusInfo>

Closest focusable in the same scope to the right of this widget.

pub fn next_up(&self) -> Option<WidgetFocusInfo>

Widget to focus when pressing the arrow up key from this widget.

pub fn next_right(&self) -> Option<WidgetFocusInfo>

Widget to focus when pressing the arrow right key from this widget.

pub fn next_down(&self) -> Option<WidgetFocusInfo>

Widget to focus when pressing the arrow down key from this widget.

pub fn next_left(&self) -> Option<WidgetFocusInfo>

Widget to focus when pressing the arrow left key from this widget.

pub fn enabled_nav(&self) -> FocusNavAction

Focus navigation actions that can move the focus away from this item.