Module zng_ext_input::gesture

Aggregate events.

Structs

ClickArgs

CLICK_EVENT arguments.

GESTURES

Gesture events config service.

GestureManager

Application extension that provides aggregate events.

ShortcutActions

Represents the resolved targets for a shortcut at a time.

ShortcutArgs

SHORTCUT_EVENT arguments.

ShortcutsHandle

Represents shortcuts claim in click_shortcut or focus_shortcut.

WeakShortcutsHandle

Weak ShortcutsHandle.

Enums

ClickArgsSource

Specific information from the source of a ClickArgs.

ShortcutClick

What kind of click a shortcut represents in a ClickArgsSource::Shortcut.

Statics

CLICK_EVENT

Aggregate click event.

SHORTCUT_EVENT

Shortcut input event.

Traits

CommandShortcutMatchesExt

Adds the shortcut_matches method to commands.

HeadlessAppGestureExt

Extension trait that adds gesture simulation methods to HeadlessApp.