zng_ext_input::gesture

Trait CommandShortcutMatchesExt

pub trait CommandShortcutMatchesExt: CommandShortcutExt {
    // Required method
    fn shortcut_matches(self, shortcut: &Shortcut) -> bool;
}

Adds the shortcut_matches method to commands.

Required Methods

fn shortcut_matches(self, shortcut: &Shortcut) -> bool

Returns true if the command has handlers, enabled or disabled, and the shortcut if one of the command shortcuts.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.