Module zng_ext_input::pointer_capture

Mouse and touch capture.

Structs

CaptureInfo

Information about mouse and touch capture in a mouse or touch event argument.

POINTER_CAPTURE

Mouse and touch capture service.

PointerCaptureArgs

POINTER_CAPTURE_EVENT arguments.

PointerCaptureManager

Application extension that provides mouse and touch capture service.

Enums

CaptureMode

Mouse and touch capture mode.

Statics

POINTER_CAPTURE_EVENT

Mouse and touch capture changed event.