zng_ext_input::touch

Struct TouchInputArgs

pub struct TouchInputArgs {

    pub timestamp: DInstant,
    pub window_id: WindowId,
    pub device_id: InputDeviceId,
    pub touch: TouchId,
    pub touch_propagation: EventPropagationHandle,
    pub position: DipPoint,
    pub force: Option<TouchForce>,
    pub velocity: DipVector,
    pub phase: TouchPhase,
    pub hits: HitTestInfo,
    pub target: InteractionPath,
    pub capture: Option<CaptureInfo>,
    pub modifiers: ModifiersState,
    pub propagation: EventPropagationHandle,

}

Arguments for TOUCH_INPUT_EVENT.

Fields
timestamp: DInstant

Instant the event happened.

window_id: WindowId

Id of window that received the event.

device_id: InputDeviceId

Id of device that generated the event.

touch: TouchId

Identify the touch contact or finger.

Multiple points of contact can happen in the same device at the same time, this ID identifies each uninterrupted contact. IDs are unique only among other concurrent touches on the same device, after a touch is ended an ID may be reused.

touch_propagation: EventPropagationHandle

Propagation handle for the touch lifetime.

The TOUCH_INPUT_EVENT and TOUCH_MOVE_EVENT have their own separate propagation handles, but touch gesture events aggregate all these events to produce a single gesture event, usually only a single gesture should be generated, multiple gestures can disambiguate using this touch_propagation handle.

As an example, TOUCH_TAP_EVENT only notifies if by the time the gesture completes the touch_propagation was not stopped. Touch gesture events or event properties must stop touch propagation as soon as they commit to a gesture, a pan gesture for example, must stop as soon as it starts scrolling, otherwise the user may accidentally scroll and tap a button at the same time.

The propagation handle always signals stopped after the touch ends. Handles are unique while at least one clone of it remains, this makes this a better unique identifier of a touch contact than TouchId that may be reused by the system as soon as a new touch contact is made.

position: DipPoint

Center of the touch in the window’s content area.

force: Option<TouchForce>

Touch pressure force and angle.

velocity: DipVector

Velocity in device independent pixels per second.

This is always zero on Start and Cancel and is the last move velocity for End. Note that the velocity value can be less than min_fling_velocity.

phase: TouchPhase

Touch phase.

Does not include Moved.

hits: HitTestInfo

Hit-test result for the touch point in the window.

target: InteractionPath

Full path to the top-most hit in hits.

capture: Option<CaptureInfo>

Current pointer capture.

modifiers: ModifiersState

What modifier keys where pressed when this event happened.

propagation: EventPropagationHandle

Propagation handle associated with this event instance.

Cloned arguments share the same handle, some arguments may also share the handle of another event if they share the same cause.

Implementations

impl TouchInputArgs

pub fn new( timestamp: impl Into<DInstant>, propagation: EventPropagationHandle, window_id: impl Into<WindowId>, device_id: impl Into<InputDeviceId>, touch: impl Into<TouchId>, touch_propagation: impl Into<EventPropagationHandle>, position: impl Into<DipPoint>, force: impl Into<Option<TouchForce>>, velocity: impl Into<DipVector>, phase: impl Into<TouchPhase>, hits: impl Into<HitTestInfo>, target: impl Into<InteractionPath>, capture: impl Into<Option<CaptureInfo>>, modifiers: impl Into<ModifiersState>, ) -> Self

New args from values that convert into the argument types.

pub fn now( window_id: impl Into<WindowId>, device_id: impl Into<InputDeviceId>, touch: impl Into<TouchId>, touch_propagation: impl Into<EventPropagationHandle>, position: impl Into<DipPoint>, force: impl Into<Option<TouchForce>>, velocity: impl Into<DipVector>, phase: impl Into<TouchPhase>, hits: impl Into<HitTestInfo>, target: impl Into<InteractionPath>, capture: impl Into<Option<CaptureInfo>>, modifiers: impl Into<ModifiersState>, ) -> Self

Arguments for event that happened now (INSTANT.now).

impl TouchInputArgs

pub fn capture_allows(&self, wgt: (WindowId, WidgetId)) -> bool

If capture is None or allows the wgt to receive this event.

pub fn is_touch_start(&self) -> bool

If the phase is start.

pub fn is_touch_end(&self) -> bool

If the phase is end.

pub fn is_touch_cancel(&self) -> bool

If the phase is cancel.

pub fn inertia_x(&self, friction: Dip) -> (Dip, Duration)

Compute the final offset and duration for a fling animation that simulates inertia movement from the velocity.x and friction. Returns 0 if velocity less than min_fling_velocity.

Friction is in dips decelerated per second.

To animate a point using these values:

Compute the final point by adding the vector offset to the current point.

Animate using the duration linear interpolation.

pub fn inertia_y(&self, friction: Dip) -> (Dip, Duration)

Compute the final offset and duration for a fling animation that simulates inertia movement from the velocity.y and friction. Returns 0 if velocity less than min_fling_velocity.

Friction is in dips decelerated per second.

pub fn position_wgt(&self, wgt: (WindowId, WidgetId)) -> Option<PxPoint>

Gets position in the widget inner bounds.