zng_ext_input::touch

Struct TouchTransformInfo

pub struct TouchTransformInfo {
    pub touches: [Point2D<f32, Px>; 2],
    pub center: Point2D<f32, Px>,
    pub deviation: f32,
    pub deviation_x: f32,
    pub deviation_y: f32,
    pub angle: AngleRadian,
}

Info useful for touch gestures computed from two touch points.

Fields
touches: [Point2D<f32, Px>; 2]

The two touch contact points.

center: Point2D<f32, Px>

Middle of the line between the two points.

deviation: f32

Deviation from the points to the center.

Min 1.0

deviation_x: f32

Deviation from the points.x to the center.x.

Min 1.0

deviation_y: f32

Deviation from the points.y to the center.y.

Min 1.0

angle: AngleRadian

Angle of the line.

Implementations

impl TouchTransformInfo

pub fn new_f32(touches: [Point2D<f32, Px>; 2]) -> Self

Compute the line info.

pub fn new(touches: [PxPoint; 2]) -> Self

Compute the line info, from round pixels.

pub fn new_dip(touches: [DipPoint; 2], scale_factor: Factor) -> Self

Compute the line info, from device independent pixels.

impl TouchTransformInfo

pub fn translation(&self, other: &Self) -> Vector2D<f32, Px>

Computes the translation to transform from self to other.

pub fn translation_x(&self, other: &Self) -> f32

Computes the translation-x to transform from self to other.

pub fn translation_y(&self, other: &Self) -> f32

Computes the translation-y to transform from self to other.

pub fn rotation(&self, other: &Self) -> AngleRadian

Computes the rotation to transform from self to other.

pub fn scale(&self, other: &Self) -> Factor

Computes the scale to transform from self to other.

pub fn scale_x(&self, other: &Self) -> Factor

Computes the scale-y to transform from self to other.

pub fn scale_y(&self, other: &Self) -> Factor

Computes the scale-y to transform from self to other.

pub fn transform(&self, other: &Self, mode: TouchTransformMode) -> PxTransform

Computes the transform from self to other.

pub fn is_single(&self) -> bool

If the transform is only a translate calculated from a single touch contact.