Module zng_ext_input::focus

Keyboard focus manager.

Modules

cmd

Commands that control focus and Command extensions.

iter

Focusable info tree iterators.

Structs

FOCUS

Keyboard focus service.

FocusChangedArgs

FOCUS_CHANGED_EVENT arguments.

FocusInfoBuilder

Builder for FocusInfo accessible in a WidgetInfoBuilder.

FocusInfoTree

A WidgetInfoTree wrapper for querying focus info out of the widget tree.

FocusManager

Application extension that manages keyboard focus.

FocusNavAction

Represents the FocusTarget actions that move focus from the current focused widget.

FocusRequest

Focus change request.

ReturnFocusChangedArgs

RETURN_FOCUS_CHANGED_EVENT arguments.

TabIndex

Widget tab navigation position within a focus scope.

WidgetFocusInfo

WidgetInfo wrapper that adds focus information for each widget.

Enums

DirectionalNav

Directional navigation configuration of a focus scope.

FocusChangedCause

The cause of a FOCUS_CHANGED_EVENT.

FocusInfo

Focus metadata associated with a widget info tree.

FocusScopeOnFocus

Behavior of a focus scope when it receives direct focus.

FocusTarget

Focus request target.

TabNav

Tab navigation configuration of a focus scope.

Statics

FOCUS_CHANGED_EVENT

Keyboard focused widget changed event.

RETURN_FOCUS_CHANGED_EVENT

Scope return focus widget changed event.

Traits

WidgetInfoFocusExt

WidgetInfo extensions that build a WidgetFocusInfo.