Module zng_ext_input::keyboard

Keyboard manager.

Structs

KEYBOARD

Keyboard service.

KeyInputArgs

Arguments for KEY_INPUT_EVENT.

KeyRepeatConfig

System settings that define the key pressed repeat.

KeyboardManager

Application extension that provides keyboard events targeting the focused widget.

ModifiersChangedArgs

Arguments for MODIFIERS_CHANGED_EVENT.

Enums

Key

Key represents the meaning of a key press.

KeyCode

Represents the location of a physical key.

KeyLocation

The location of the key on the keyboard.

KeyState

State a Key has entered.

NativeKeyCode

Contains the platform-native physical key identifier

Statics

KEY_INPUT_EVENT

Key pressed, repeat pressed or released event.

MODIFIERS_CHANGED_EVENT

Modifiers key state changed event.

Traits

HeadlessAppKeyboardExt

Extension trait that adds keyboard simulation methods to HeadlessApp.