Module zng_ext_input::mouse

Mouse events and service.

The app extension MouseManager provides the events and service. It is included in the default application.

Structs

ButtonRepeatConfig

Settings that define the mouse button pressed repeat.

ClickMode

Defines how click events are generated for a widget.

MOUSE

Mouse service.

MouseClickArgs

MOUSE_CLICK_EVENT arguments.

MouseHoverArgs

MOUSE_HOVERED_EVENT arguments.

MouseInputArgs

MOUSE_INPUT_EVENT arguments.

MouseManager

Application extension that provides mouse events and service.

MouseMoveArgs

MOUSE_MOVE_EVENT arguments.

MousePosition

Mouse cursor position.

MouseWheelArgs

MOUSE_WHEEL_EVENT arguments.

MultiClickConfig

System settings needed for implementing double/triple clicks.

Enums

ButtonState

State a MouseButton has entered.

ClickTrigger

Represents mouse gestures that can initiate a click.

MouseButton

Describes a button of a mouse controller.

MouseScrollDelta

Describes a difference in the mouse scroll wheel state.

Statics

MOUSE_CLICK_EVENT

Mouse click event, any click_count.

MOUSE_HOVERED_EVENT

The top-most hovered widget changed or pointer capture changed.

MOUSE_INPUT_EVENT

Mouse down or up event.

MOUSE_MOVE_EVENT

Mouse move event.

MOUSE_WHEEL_EVENT

Mouse wheel scroll event.

Traits

WidgetInfoBuilderMouseExt

Mouse config builder methods.

WidgetInfoMouseExt

Mouse config methods.