zng_ext_input
Module touch

Touch events and service.

The app extension TouchManager provides the events and service. It is included in the default application.

Structs

TOUCH

Touch service.

TouchConfig

System settings needed to implementing touch gestures.

TouchId

Identifier for a continuous touch contact.

TouchInputArgs

Arguments for TOUCH_INPUT_EVENT.

TouchLongPressArgs

Arguments for TOUCH_LONG_PRESS_EVENT.

TouchManager

Application extension that provides touch events and service.

TouchMove

Identify the moves of one touch contact in TouchMoveArgs.

TouchMoveArgs

Arguments for TOUCH_MOVE_EVENT.

TouchPosition

Active touch positions.

TouchTapArgs

Arguments for TOUCH_TAP_EVENT.

TouchTransformArgs

Arguments for TOUCH_TRANSFORM_EVENT.

TouchTransformInfo

Info useful for touch gestures computed from two touch points.

TouchTransformMode

Defines the different transforms that a touch transform can do to keep two touch points in a widget aligned with the touch contacts.

TouchUpdate

Identify a new touch contact or a contact update.

TouchedArgs

Arguments for TOUCHED_EVENT.

Enums

TouchForce

Describes the force of a touch event.

TouchPhase

Describes touch-screen input state.

Statics

TOUCHED_EVENT

Touch made first contact or lost contact with a widget.

TOUCH_INPUT_EVENT

Touch contact started or ended.

TOUCH_LONG_PRESS_EVENT

Touch contact pressed without moving for more then the tap_max_time.

TOUCH_MOVE_EVENT

Touch contact moved.

TOUCH_TAP_EVENT

Touch tap.

TOUCH_TRANSFORM_EVENT

Two point touch transform.